# backend/src/swagger-output.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Newsx API",
    "description": "API Documentation for Newsx application, a platform for creating and managing blog posts, comments, likes, and user interactions.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://blog-sphere-backend-ruby.vercel.app",
      "description": "Production server"
    }
  ],
  "basePath": "/",
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter token in format: Bearer <token>"
    }
  },
  "definitions": {
    "ApiResponse": {
      "type": "object",
      "properties": {
        "statusCode": { "type": "integer", "example": 200 },
        "data": { "type": "object", "nullable": true },
        "message": { "type": "string", "example": "Operation successful" },
        "success": { "type": "boolean", "example": true }
      }
    },
...
  },
  "paths": {
...
    "/api/v1/users/{userName}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a single user and their posts",
        "description": "Retrieves a user’s profile and posts by their username.",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "userName",
            "description": "Username of the user",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User and posts fetched",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/ApiResponse" },
                "example": {
                  "statusCode": 200,
                  "data": {
                    "_id": "60d21b4667d0d8992e610c84",
                    "name": "John Doe",
                    "email": "john@example.com",
                    "userName": "johndoe",
                    "posts": [
                      {
                        "_id": "60d21b4667d0d8992e610c85",
                        "title": "My First Post"
                      }
                    ]
                  },
                  "message": "User and posts fetched successfully",
                  "success": true
                }
              }
            }
          },
          "404": { "description": "User not found" }
        }
      }
    },
...
  },
...
}